+++ concat.go
@@ -1,0 +1,22 @@
+package datatable
+
+// func Concat(tables ...*DataTable) *DataTable {
+// 	out := New("???")
+
+// 	cols := make(map[string]int)
+
+// 	for i, t := range tables {
+
+// 		// Grow !
+
+// 		for j, c := range t.cols {
+// 			if col, ok := cols[c.name]; ok {
+
+// 			}
+
+// 		}
+
+// 	}
+
+// 	return out
+// }
