--- datatable/export.go
@@ -1,16 +1,0 @@
-package datatable
-
-type ExportOptions struct {
-	HiddenColumns  []string
-	IncludeHeaders bool
-}
-
-type ExportOption func(opts *ExportOptions)
-
-func newExportOptions(opt ...ExportOption) ExportOptions {
-	options := ExportOptions{}
-	for _, o := range opt {
-		o(&options)
-	}
-	return options
-}
