+++ export/slice.go
