+++ hasher.go
