+++ mutate_row_test.go
@@ -3,11 +3,11 @@ package datatable_test
 import (
 	"testing"
 
-	"github.com/datasweet/datatable/datatable"
+	"github.com/datasweet/datatable"
 	"github.com/datasweet/datatable/serie"
 )
 
-func TestSwapColumn(t *testing.T) {
+func TestSwapRow(t *testing.T) {
 	tb := datatable.New("test")
 	tb.AddColumn("champ", serie.String("Malzahar", "Xerath", "Teemo"))
 	tb.AddExprColumn("champion", serie.String(), "upper(`champ`)")
@@ -21,12 +21,12 @@ func TestSwapColumn(t *testing.T) {
 		"Teemo", "TEEMO", 666, 666, 50.0,
 	)
 
-	tb.SwapColumn("champion", "winRate")
+	tb.SwapRow(0, 2)
 
 	checkTable(t, tb,
-		"champ", "winRate", "win", "loose", "champion",
-		"Malzahar", 62.5, 10, 6, "MALZAHAR",
-		"Xerath", 80.0, 20, 5, "XERATH",
-		"Teemo", 50.0, 666, 666, "TEEMO",
+		"champ", "champion", "win", "loose", "winRate",
+		"Teemo", "TEEMO", 666, 666, 50.0,
+		"Xerath", "XERATH", 20, 5, 80.0,
+		"Malzahar", "MALZAHAR", 10, 6, 62.5,
 	)
 }
